Add Users namespace
Controllers which require authentication moved to `users` namespace.

Added ability to search views without namespace/prefix.

When we render `#new` action renders default view then `articles/new` will be selected
for render.

Checkout `BaseController#local_prefixes`
and `config.action_view.prefix_partial_path_with_controller_namespace` option
in `config/application.rb`
