+++ app/models/comment.rb
@@ -1,4 +1,6 @@
 class Comment < ActiveRecord::Base
   belongs_to :article
   validates :text, presence: true
+
+  scope :by_article, ->(article) { where(article: article) }
 end
