+++ app/views/application/_navigation_main.html.slim
@@ -2,4 +2,4 @@ ul.left
   - if user_signed_in?
     = active_link_to "Home", root_path, active: :exclusive, wrap_tag: :li
     = active_link_to "Articles", articles_path, active: :exclusive, wrap_tag: :li
-    = active_link_to "Create Article", new_user_article_path, active: :exclusive, wrap_tag: :li
+    = active_link_to "Create Article", new_users_article_path, active: :exclusive, wrap_tag: :li
