+++ app/views/articles/_form.html.slim
@@ -1,6 +1,6 @@
 .row
   .medium-6.columns
-    = simple_form_for([:user, article]) do |f|
+    = simple_form_for([:users, article]) do |f|
       .form-inputs
         = f.input :title
         = f.input :text
