+++ app/views/articles/show.html.slim
@@ -1,6 +1,6 @@
 .container
   = render(article)
 
-  = link_to("Edit", [:edit, :user, article]) if policy(article).manage?
+  = link_to("Edit", [:edit, :users, article]) if policy(article).manage?
 
 = render("comments/discussion", comments: comments, comment: comment)
