+++ app/views/comments/_form.html.slim
@@ -1,6 +1,6 @@
 .row
   .medium-6.columns
-    = simple_form_for([comment.article, :user, comment], data: { remote: true, behavior: "fragments" }) do |f|
+    = simple_form_for([:users, comment.article, comment], data: { remote: true, behavior: "fragments" }) do |f|
       .form-inputs
         = f.input :text
