+++ config/rails_best_practices.yml
@@ -1,6 +1,6 @@
 AddModelVirtualAttributeCheck: { }
 AlwaysAddDbIndexCheck: { }
-CheckSaveReturnValueCheck: { }
+# CheckSaveReturnValueCheck: { }
 DefaultScopeIsEvilCheck: { }
 DryBundlerInCapistranoCheck: { }
 HashSyntaxCheck: { }
@@ -25,11 +25,9 @@ ProtectMassAssignmentCheck: { }
 RemoveEmptyHelpersCheck: { }
 RemoveTabCheck: { }
 RemoveTrailingWhitespaceCheck: { }
-RemoveUnusedMethodsInControllersCheck:
-  except_methods:
-  - ApplicationController#devise_parameter_sanitizer
-RemoveUnusedMethodsInHelpersCheck: { except_methods: [] }
-RemoveUnusedMethodsInModelsCheck: { except_methods: [] }
+# RemoveUnusedMethodsInControllersCheck: { except_methods: [] }
+# RemoveUnusedMethodsInHelpersCheck: { except_methods: [] }
+# RemoveUnusedMethodsInModelsCheck: { except_methods: [] }
 ReplaceComplexCreationWithFactoryMethodCheck: { attribute_assignment_count: 2 }
 ReplaceInstanceVariableWithLocalVariableCheck: { }
 RestrictAutoGeneratedRoutesCheck: { }
